+++ src/ZPublisher/HTTPResponse.py
@@ -264,11 +264,15 @@ class HTTPBaseResponse(BaseResponse):
 
         This value overwrites any previously set value for the
         cookie in the Response object.
+
+        `name` has to be text in Python 3.
+
+        `value` may be text or bytes. The default encoding of respective python
+        version is used.
         """
-        if PY2 and isinstance(name, text_type):
-            name = name.encode(self.charset)
-        if isinstance(value, text_type):
-            value = value.encode(self.charset)
+        if PY2:
+            name = str(name)
+            value = str(value)
 
         cookies = self.cookies
         if name in cookies:
@@ -287,11 +291,15 @@ class HTTPBaseResponse(BaseResponse):
         browsers with a key "name" and value "value". If a value for the
         cookie has previously been set in the response object, the new
         value is appended to the old one separated by a colon.
+
+        `name` has to be text in Python 3.
+
+        `value` may be text or bytes. The default encoding of respective python
+        version is used.
         """
-        if PY2 and isinstance(name, text_type):
-            name = name.encode(self.charset)
-        if isinstance(value, text_type):
-            value = value.encode(self.charset)
+        if PY2:
+            name = str(name)
+            value = str(value)
 
         cookies = self.cookies
         if name in cookies:
@@ -313,9 +321,11 @@ class HTTPBaseResponse(BaseResponse):
         to be specified - this path must exactly match the path given
         when creating the cookie. The path can be specified as a keyword
         argument.
+
+        `name` has to be text in Python 3.
         """
-        if PY2 and isinstance(name, text_type):
-            name = name.encode(self.charset)
+        if PY2:
+            name = str(name)
 
         d = kw.copy()
         if 'value' in d: